feat: replace beautysh with fmt and add markdownlint as dev package
Update mason and formatting neovim configuration to replace beautysh
with shfmt as the formatter for shell scripts.

Introduce markdownlint-cli as a dev package in install_functions.sh and
replace beautysh with shfmt as a dev package.

Update utils.sh file to declare and assign border local var separately
in header() function to avoid masking return values.