feat(gtk): add qt wayland packages, gtk theme and gtk config
Introduce orchis-theme, qt5-wayland and qt6-wayland as required packages
in install_functions.sh. Introduce GTK configuration files to set the
theme, cursor and font for GTK applications.

Update hyprland config file to change cursor and waybar config file to
adjust the padding of the clock module and margin of the tray module.